+++ src/main/java/repast/model/agrosuccess/AgroSuccessContextBuilder.java
@@ -6,6 +6,7 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Set;
 import org.apache.commons.configuration.ConfigurationException;
 import org.apache.log4j.Logger;
 import org.neo4j.graphdb.GraphDatabaseService;
@@ -39,6 +40,8 @@ import me.ajlane.geo.repast.succession.GraphBasedLcsTransitionMapFactory;
 import me.ajlane.geo.repast.succession.LcsTransitionMapFactory;
 import me.ajlane.geo.repast.succession.LcsUpdateDecider;
 import me.ajlane.geo.repast.succession.LcsUpdater;
+import me.ajlane.geo.repast.succession.SeedStateUpdater;
+import me.ajlane.geo.repast.succession.SuccessionPathwayUpdater;
 import me.ajlane.neo4j.EmbeddedGraphInstance;
 import repast.model.agrosuccess.AgroSuccessCodeAliases.Lct;
 import repast.model.agrosuccess.empirical.SiteAllData;
@@ -174,6 +177,8 @@ public class AgroSuccessContextBuilder implements ContextBuilder<Object> {
     layerList.add(uniformDefaultLayer(LscapeLayer.DeltaD, -1, gridDimensions, gridOrigin));
     layerList.add(uniformDefaultLayer(LscapeLayer.DeltaT, -1, gridDimensions, gridOrigin));
     layerList.add(uniformDefaultLayer(LscapeLayer.TimeInState, 0, gridDimensions, gridOrigin));
+    layerList.add(uniformDefaultLayer(LscapeLayer.FireCount, 0, gridDimensions, gridOrigin));
+    layerList.add(uniformDefaultLayer(LscapeLayer.OakAge, -1, gridDimensions, gridOrigin));
 
     try {
       layerList.add(siteRasterData.getLctMap());
@@ -262,9 +267,15 @@ public class AgroSuccessContextBuilder implements ContextBuilder<Object> {
     LcsTransitionMapFactory fac =
         new GraphBasedLcsTransitionMapFactory(graph, graphModelID, translator);
     CodedLcsTransitionMap codedMap = fac.getCodedLcsTransitionMap();
+
+    // TODO Refactor so oak mortality scaling parameter loaded from graph
+    SuccessionPathwayUpdater successionUpdater = new SuccessionPathwayUpdater(200.);
     // TODO Refactor so set of mature land-cover types loaded from graph
-    LcsUpdateDecider updateDecider = new AgroSuccessLcsUpdateDecider(codedMap, new HashSet<Integer>(
-        Arrays.asList(Lct.Pine.getCode(), Lct.Oak.getCode(), Lct.Deciduous.getCode())));
+    SeedStateUpdater seedUpdater = new SeedStateUpdater(
+        new HashSet<Integer>(Arrays.asList(Lct.Pine.getCode(), Lct.Oak.getCode(), Lct.Deciduous
+            .getCode())));
+    LcsUpdateDecider updateDecider = new AgroSuccessLcsUpdateDecider(codedMap, successionUpdater,
+        seedUpdater);
 
     SoilMoistureDiscretiser smDiscretiser = new AgroSuccessSoilMoistureDiscretiser(smParams);